Created initial classes and methods.
Problems - this class will hold all of the problems as methods of their own.

DynamicProblemPrinter - By using reflection, we can call any problem from the problems class which in turn will be shown to the screen.

Program- I added some of the methods that will be use to interact with the user.