Fixed small issues in DynamicProblemPrinter.
Changed the name of the class to DynamicProblemSelector since this name is more appropriate.

The class is now public static since it is the most logical approach.

Wrapped the method.Invoke in an if statement that checks for null to avoid any disasters in case user enters a problem number we do not have or is simply incorrect.